requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client
langchain
flask
pypdf
pinecone-client
python-dotenv
setuptools